fix: move out methods that don't belong to helpers
