+++ app/controllers/application_controller.rb
@@ -1,3 +1,6 @@
 class ApplicationController < ActionController::Base
-  include SessionsHelper
+  include Authorization
+  include Authentication
+
+  helper_method :logged_in?, :current_user
 end
