+++ app/helpers/sessions_helper.rb
@@ -1,58 +1,2 @@
 module SessionsHelper
-  def log_in(user)
-    session[:user_id] = user.id
-  end
-
-  def log_out
-    forget(current_user)
-    session.delete(:user_id)
-    @current_user = nil
-  end
-
-  def remember(user)
-    user.remember
-    cookies.permanent.signed[:user_id] = user.id
-    cookies.permanent[:remember_token] = user.remember_token
-  end
-
-  def forget(user)
-    user.forget
-    cookies.delete(:user_id)
-    cookies.delete(:remember_token)
-  end
-
-  def current_user
-    current_cookies || current_session
-  end
-
-  def current_session 
-    if (user_id = session[:user_id])
-      @current_user ||= User.find(user_id) 
-    end
-  end
-
-  def current_cookies 
-    if (user_id = cookies.signed[:user_id])
-      authenticated?(User.find(user_id)) 
-    end
-  end
-
-  def authenticated?(user)
-    if user && user.authenticated?(cookies[:remember_token])
-      log_in user
-      @current_user = user
-    end
-  end
-    
-
-  def logged_in?
-    !!current_user
-  end
-
-  def require_logout
-    if logged_in?
-      flash[:error] = "You are already logged in"
-      redirect_to me_url
-    end
-  end
 end
